requirements.txt
Django
djangorestframework
dnspython
djangorestframework-simplejwt
django-cors-headers
drf-yasg
django-heroku
PyJWT
mysqlclient
gunicorn